requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.4
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-files==1.0.1
drf-pdf==0.2.0
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
openapi-codec==1.3.2
pbr==3.1.1
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
